Extract tests for isPromise to own file
